This is a copy of the commit message I was going to use for the other branch (the Anand branch), instead I'm submitting the same files to the main branch.
Also note that on top of the written log below I also now have insert
queries working. Also the cont_net program sends back 1 of three
messages depending on context (SUCCESS|OUTPUT|ERROR) which the Client
java program must interpret.

Making MAJOR changes to the DistributedDB class and operation
as of this writing, the distributedDB.Client program
launches a container on all found Nodes. The Client program
binds to port "12345" by default and so it is accessible via a
telnet NODENAME 12345 command.

The AppMaster and all
container nodes register themselves back to the Client, and the
Client uses a TCPServer class to hold onto the Channel for each
of these nodes so that we can communicate directly back to that node.
This eliminates the need for generating a new TCPClient connection
and requiring that each node advertises a server port.

The Containers launch a new Python program called cont_net.py
This program replaces the ContainerService class from java.
Also now there is no bash script wrapper, the java code directly
calls cont_net.py.

cont_net.py registers itself to the Client, and waits for queries.
When it receives a new query it generates a sqlite binary database
name in the form of "out-APPID-NODE-QUERYNUMBER" and calls the exec_cmd.py program.
This program takes any output that the script would return (as part
of a select statement) and stores it in a new table which has a schema
identical to the schema from the select table. The exec_cmd.py program
then calls an hdfs command which stores the output sqlite database into
HDFS. Finally, exec_cmd.py outputs either "success" (if an insert or
create query) or "out-APPID-NODENAME" (if a select query).
This data is captured by cont_net.py and transmitted back to the Client
program.

Once the Client has the response from cont_net.py, it searches for a
file in HDFS titled "out-APPID-NODENAME" and copies it to the /tmp
directory. From there it calls a bash script I created called
gather_sqlite_results.sh. This script takes in the sql statement
required to extract data from the table, and a list of all database
image names that were generated by the different nodes, it then
generates the proper sqlite3 commands and executes them, thus gathering
all data from the tables and concating it into one output, which is
finally sent back to the user.

PLEASE NOTE:
1) Not all of the scripts required are copied around by the YARN
framework so it would be best to copy all files from the scripts folder
to the home directory of every node in the hadoop cluster (to
/home/hduser).