Adding the ability to perform processing in parallel and also adding the ability to print the elapsed time of the query. These are both 'menu' options which can be seen by typing '!help'.
Also I forgot to mention a MAJOR CAVEAT with my last commit:

1) ONLY THE RANDOM PARTITION SCHEME is FULLY IMPLEMENTED right now
so the only valid create command is:
"creat table test(stuff) partition by random(x)"

The "x" is optional and it specifies how many nodes of the system to
use. If it isn't specified (the query should look like "partition by
random" then NOT "random()") then the full number of nodes discovered is
used.