Adding quickstep support scripts
I'm adding the initial quickstep scripts that are executed by the Yarn
client and the Yarn containers.

exec_qs.py is a persistant python program that acts just like
cont_net.py except that it starts up a quickstep process and then
continues to run, it interfaces with the Yarn client and forwards
commands to the quickstep CLI component.

gather_qs.py is a program that runs on the Client node, it is executed
by the Yarn Client and is a replacement of the gather_sqlite_results.sh
script. It expects to get a copy of block names which have been moved
from HDFS into /tmp/ on that computer, as well as a JSON that represents
the schema definition for the table that has been requested.