Massive commit for random fixes.
I changed the quickstep_cli_shell binary to prepend the string "shell"
to the beginning of all of its internal data storage and catalog files
so that on nodes where the Client and a slave are both running they
won't collide. I also fixed gather_qs to deal with that new change.

I fixed a wierd Log error in Partition java code and a few typoes.

I fixed quickstep_cli_net to perform a sockopt to set the REUSEADDR flag
on the TCP server socket so it should stop failing randomly.

I added logic to the Yarn Client so that while using quickstep if a node
returns "SUCCESS" rather than an "OUTPUT" statement for a select query,
it will not print that out (I think this is an OK situation) but I also
added logic so that if ALL of the nodes returned "SUCCESS" then it does
print an error message because that is an error (I guess unless the
table was actually empty).

I added logic to the DDBPartitioner code so that it saves a different
partitions.data file depending on the type of DB you use, it will either
be "quicksteppartitions.data" or "sqlite3parititions.data" now.